requirements.txt
torchvision
torchaudio
numpy
face_alignment
imageio
imageio-ffmpeg
librosa
numba
resampy
pydub 
scipy
kornia
tqdm
yacs
pyyaml  
joblib
scikit-image
basicsr
facexlib
dlib-bin
gfpgan
flytekit
